Use Go's built-in fuzzing and publish a Homebrew tap
The targets in `fuzz/` carried a `gofuzz` build tag for dvyukov/go-fuzz,
so the toolchain excluded the whole package and never compiled, vetted
or tested it. They are now `testing.F` targets in `_test.go` files with
seed corpora, which `go test ./...` runs, and `make fuzz` drives
`go test -fuzz` instead of `go-fuzz-build`.

The Homebrew cask targets kjanat/homebrew-actionlint instead of
rhysd/actionlint, and `skip_upload` is gone so every release updates it.
GoReleaser needs a `HOMEBREW_TAP_TOKEN` secret for that push because the
built-in `GITHUB_TOKEN` cannot write to another repository.

The action examples in docs/usage.md now use the `v1` tag.
