+++ .goreleaser.yaml
@@ -62,12 +62,12 @@ homebrew_casks:
     dependencies:
       - formula: shellcheck
     repository:
-      owner: rhysd
-      name: actionlint
+      owner: kjanat
+      name: homebrew-actionlint
+      token: "{{ .Env.HOMEBREW_TAP_TOKEN }}"
     commit_author:
       name: "github-actions[bot]"
       email: "41898282+github-actions[bot]@users.noreply.github.com"
-    skip_upload: true
 
 changelog:
   disable: true
