+++ fuzz/README.md
@@ -1,1 +1,3 @@
 See [CONTRIBUTING.md](../CONTRIBUTING.md) about this directory.
+
+<!-- markdownlint-disable-file first-line-heading -->
