--- fuzz/glob.go
@@ -1,23 +1,0 @@
-//go:build gofuzz
-
-package actionlint_fuzz
-
-import (
-	"github.com/kjanat/actionlint"
-)
-
-func FuzzGlobGitRef(data []byte) int {
-	errs := actionlint.ValidateRefGlob(string(data))
-	if len(errs) > 0 {
-		return 0
-	}
-	return 1
-}
-
-func FuzzGlobFilePath(data []byte) int {
-	errs := actionlint.ValidatePathGlob(string(data))
-	if len(errs) > 0 {
-		return 0
-	}
-	return 1
-}
